requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
transformers==4.35.0
torch==2.1.0
spacy==3.7.2
scikit-learn==1.3.2
python-dotenv==1.0.0
redis==5.0.1
numpy==1.24.3
newsapi-python==0.2.7
aiohttp==3.9.1
python-dateutil==2.8.2 